server/requirement.txt
Django
django-cors-headers
djangorestframework
Pillow
django-rest-knox